[TASK] Move all Slot functionality directly into Event Listener
The outdated naming was confusing
